+++ app/src/background_description.js
@@ -1,13 +1,15 @@
 // Constants
 const DESCRIPTION_SELECTOR =
-  "#description-inline-expander, ytd-expander#description, .expandable-video-description-body-main, .expandable-video-description-container, #collapsed-string, #expanded-string";
+  "#description-inline-expander, ytd-expander#description, .expandable-video-description-body-main, .expandable-video-description-container, #collapsed-string, #expanded-string, #anchored-panel ytd-text-inline-expander";
 const AUTHOR_SELECTOR = `#upload-info.ytd-video-owner-renderer, 
-div.slim-owner-bylines > h3.slim-owner-channel-name,  
-div.cbox > a > h3.reel-player-header-channel-title`;
-const ATTRIBUTED_STRING_SELECTOR =
-  "yt-attributed-string, .yt-core-attributed-string";
+ytm-slim-owner-renderer div.slim-owner-bylines,  
+div.cbox > a.reel-player-header-channel-endpoint.cbox`;
+const ATTRIBUTED_STRING_SELECTOR = "yt-attributed-string";
+
+const ATTRIBUTED_STRING_CLASS_SELECTOR = ".yt-core-attributed-string";
 const FORMATTED_STRING_SELECTOR = "yt-formatted-string";
-const SNIPPET_TEXT_SELECTOR = "#attributed-snippet-text";
+const SNIPPET_TEXT_SELECTOR =
+  "#attributed-snippet-text, #formatted-snippet-text, #plain-snippet-text";
 const HORIZONTAL_CHAPTERS_SELECTOR =
   "ytd-horizontal-card-list-renderer, ytd-macro-markers-list-renderer";
 const CHAPTER_ITEM_SELECTOR = "ytd-macro-markers-list-item-renderer";
@@ -472,8 +474,10 @@ function setupChapters(originalDescription) {
     });
 
     if (shouldUpdate) {
-      // Debounce updates
-      setTimeout(updateTooltipChapter, 16); // ~60fps instead of immediate
+      // YouTube updates tooltip text on every mouse move along timeline [#132](https://github.com/zpix1/yt-anti-translate/issues/132)
+      // debounce using requestAnimationFrame because setTimeout is too slow and no debounce at all is laggy
+      // requestAnimationFrame only works with tab visible but that's acceptable for this use case as user won't be able to use the tooltip when tab is not visible
+      requestAnimationFrame(updateTooltipChapter);
     }
   });
 
@@ -562,7 +566,7 @@ function getPlayerResponseSafely(playerEl) {
  *
  * @returns {string|null} The original description or null if it cannot be retrieved.
  */
-function fetchOriginalDescription() {
+async function fetchOriginalDescription() {
   const player = window.YoutubeAntiTranslate.getFirstVisible(
     document.querySelectorAll(window.YoutubeAntiTranslate.getPlayerSelector()),
   );
@@ -570,14 +574,18 @@ function fetchOriginalDescription() {
   const playerResponse = getPlayerResponseSafely(player);
   if (!playerResponse && window.YoutubeAntiTranslate.isMobile()) {
     // Fallback for mobile layout when player response is unavailable
-    const mobileDescription = getDescriptionMobile();
+    const mobileDescription = await getDescriptionMobile();
     if (mobileDescription) {
       return mobileDescription;
     }
     return null;
   }
 
-  return playerResponse?.videoDetails?.shortDescription || null;
+  return (
+    playerResponse?.videoDetails?.shortDescription ||
+    playerResponse?.videoDetails?.title ||
+    null
+  );
 }
 
 /**
@@ -607,8 +615,8 @@ function fetchOriginalAuthor() {
  * Restores the original (untranslated) description and author name in the DOM and
  * triggers the chapters replacement logic.
  */
-function restoreOriginalDescriptionAndAuthor() {
-  const originalDescription = fetchOriginalDescription();
+async function restoreOriginalDescriptionAndAuthor() {
+  const originalDescription = await fetchOriginalDescription();
   const originalAuthor = fetchOriginalAuthor();
 
   if (!originalDescription && !originalAuthor) {
@@ -682,7 +690,7 @@ function updateDescriptionContent(container, originalText) {
   // Find the text containers
   const mainTextContainer = window.YoutubeAntiTranslate.getFirstVisible(
     container.querySelectorAll(
-      `${ATTRIBUTED_STRING_SELECTOR}, ${FORMATTED_STRING_SELECTOR}`,
+      `${ATTRIBUTED_STRING_SELECTOR}, ${ATTRIBUTED_STRING_CLASS_SELECTOR}, ${FORMATTED_STRING_SELECTOR}`,
     ),
   );
   const snippetTextContainer = window.YoutubeAntiTranslate.getFirstVisible(
@@ -728,7 +736,16 @@ function updateDescriptionContent(container, originalText) {
 
   // Check each container independently
   const mainNeedsUpdate = mainTextContainer
-    ? needsUpdate(mainTextContainer)
+    ? !mainTextContainer.closest(
+        SNIPPET_TEXT_SELECTOR,
+      ) /*mainTextContainer selector can include the children of snippetTextContainer so make sure that snippetTextContainer is not a parent*/ &&
+      !mainTextContainer.querySelector(
+        "#description-placeholder",
+      ) /*ignore placeholder*/ &&
+      !mainTextContainer.querySelector(
+        '[style="color: rgb(170, 170, 170);"',
+      ) /*ignore grey text*/ &&
+      needsUpdate(mainTextContainer)
     : false;
   const snippetNeedsUpdate = snippetTextContainer
     ? needsUpdate(snippetTextContainer)
@@ -769,16 +786,29 @@ function updateDescriptionContent(container, originalText) {
  */
 function updateAuthorContent(container, originalText) {
   // Find the text containers
-  const mainTextContainer = window.YoutubeAntiTranslate.getFirstVisible(
-    container.querySelectorAll(FORMATTED_STRING_SELECTOR),
-  );
-  const snippetTextContainer = window.YoutubeAntiTranslate.getFirstVisible(
-    container.querySelectorAll(
-      `${FORMATTED_STRING_SELECTOR} a, ${ATTRIBUTED_STRING_SELECTOR}`,
-    ),
-  );
+  const singularChannelNameTitleContainer =
+    window.YoutubeAntiTranslate.getFirstVisible(
+      container.querySelectorAll(`#channel-name ${FORMATTED_STRING_SELECTOR}`),
+    );
+  const singularChannelNameTextContainer =
+    window.YoutubeAntiTranslate.getFirstVisible(
+      container.querySelectorAll(
+        `#channel-name ${FORMATTED_STRING_SELECTOR} a, #channel-name ${ATTRIBUTED_STRING_SELECTOR}, #channel-name ${ATTRIBUTED_STRING_CLASS_SELECTOR}, .slim-owner-channel-name > ${ATTRIBUTED_STRING_CLASS_SELECTOR}, 
+        .reel-player-header-channel-title > ${ATTRIBUTED_STRING_CLASS_SELECTOR}`,
+      ),
+    );
+  const multipleChannelNameContainers =
+    window.YoutubeAntiTranslate.getFirstVisible(
+      container.querySelectorAll(
+        `#attributed-channel-name ${ATTRIBUTED_STRING_CLASS_SELECTOR} a > span`,
+      ),
+    );
 
-  if (!mainTextContainer && !snippetTextContainer) {
+  if (
+    !singularChannelNameTitleContainer &&
+    !singularChannelNameTextContainer &&
+    !multipleChannelNameContainers
+  ) {
     window.YoutubeAntiTranslate.logWarning(
       `No video author text containers found`,
     );
@@ -786,22 +816,32 @@ function updateAuthorContent(container, originalText) {
   }
 
   // Update both containers if they exist
-  if (mainTextContainer) {
-    if (mainTextContainer.title !== originalText) {
-      mainTextContainer.title = originalText;
+  if (singularChannelNameTitleContainer) {
+    if (singularChannelNameTitleContainer.title !== originalText) {
+      singularChannelNameTitleContainer.title = originalText;
     }
   }
 
-  if (snippetTextContainer) {
-    if (snippetTextContainer.innerText !== originalText) {
+  if (singularChannelNameTextContainer) {
+    if (singularChannelNameTextContainer.innerText !== originalText) {
       const storeStyleDisplay =
-        snippetTextContainer.parentElement.style.display;
-      snippetTextContainer.parentElement.style.display = "none";
-      snippetTextContainer.innerText = originalText;
+        singularChannelNameTextContainer.parentElement.style.display;
+      singularChannelNameTextContainer.parentElement.style.display = "none";
+      singularChannelNameTextContainer.innerText = originalText;
       // Force reflow
       setTimeout(() => {
-        snippetTextContainer.parentElement.style.display = storeStyleDisplay;
+        singularChannelNameTextContainer.parentElement.style.display =
+          storeStyleDisplay;
       }, 50);
+    }
+  }
+
+  if (multipleChannelNameContainers) {
+    const firstTextNode = window.YoutubeAntiTranslate.getFirstTextNode(
+      multipleChannelNameContainers,
+    );
+    if (firstTextNode && firstTextNode.textContent !== originalText) {
+      firstTextNode.textContent = originalText;
     }
   }
 }
@@ -814,7 +854,7 @@ async function handleDescriptionMutation() {
     document.querySelectorAll(window.YoutubeAntiTranslate.getPlayerSelector()),
   );
   if (descriptionElement && player) {
-    restoreOriginalDescriptionAndAuthor();
+    await restoreOriginalDescriptionAndAuthor();
   }
 
   // On mobile the author is visible even when the description is not
@@ -1031,9 +1071,46 @@ function extractVideoDataField(fieldName) {
   }
 }
 
+// Get Title using oembed for mobile
+async function getTitleMobile(url) {
+  const videoId = window.YoutubeAntiTranslate.extractVideoIdFromUrl(url);
+  if (!videoId) {
+    return null;
+  }
+
+  let response = await window.YoutubeAntiTranslate.cachedRequest(
+    `https://www.youtube.com/oembed?url=https://www.youtube.com/watch?v=${videoId}`,
+  );
+  if (
+    !response ||
+    !response.response ||
+    !response.response.ok ||
+    !response.data?.title
+  ) {
+    if (response?.response?.status === 401) {
+      // 401 likely means the video is restricted try again with youtubeI
+      response =
+        await window.YoutubeAntiTranslate.getVideoTitleFromYoutubeI(videoId);
+      if (!response?.response?.ok || !response.data?.title) {
+        window.YoutubeAntiTranslate.logWarning(
+          `YoutubeI title request failed for video ${videoId}`,
+        );
+        return;
+      }
+    } else {
+      return;
+    }
+  }
+
+  return response.data.title;
+}
+
 // Get Description from the VideoData
-function getDescriptionMobile() {
-  return extractVideoDataField("shortDescription");
+async function getDescriptionMobile() {
+  return (
+    extractVideoDataField("shortDescription") ||
+    (await getTitleMobile(document.location.href))
+  );
 }
 
 // Get Author from the VideoData