+++ tests/youtube-extension.spec.ts
@@ -1,3 +1,5 @@
+/* eslint-disable  @typescript-eslint/no-explicit-any */
+
 import { expect } from "@playwright/test";
 import { test } from "../playwright.config";
 import {
@@ -139,6 +141,66 @@ test.describe("YouTube Anti-Translate extension", () => {
     // Take a screenshot for visual verification
     await page.screenshot({
       path: `images/tests/${browserNameWithExtensions}/${localeString}/youtube-extension-test-description.png`,
+    });
+
+    // Check console message count
+    expect(consoleMessageCountContainer.count).toBeLessThan(2000);
+
+    // Close the browser context
+    await context.close();
+  });
+
+  test("Prevents current video description to fallback to translated title", async ({
+    browserNameWithExtensions,
+    localeString,
+  }, testInfo) => {
+    await handleRetrySetup(testInfo, browserNameWithExtensions, localeString);
+
+    // Launch browser with the extension
+    const context = await createBrowserContext(browserNameWithExtensions);
+
+    // Create a new page
+    const { page, consoleMessageCountContainer } = await setupPageWithAuth(
+      context,
+      browserNameWithExtensions,
+      localeString,
+    );
+
+    await loadPageAndVerifyAuth(
+      page,
+      "https://www.youtube.com/watch?v=50G0kIty7Cg",
+      browserNameWithExtensions,
+    );
+
+    // Wait for the video page to fully load
+    await page.waitForSelector("ytd-watch-metadata");
+
+    // Expand the description if it's collapsed
+    const moreButton = page.locator(
+      "#description-inline-expander ytd-text-inline-expander #expand",
+    );
+    if (await moreButton.isVisible()) {
+      await moreButton.click();
+      // Wait for the description to expand
+      await page.waitForTimeout(1000);
+    }
+
+    // Get the description text
+    const descriptionText = await page
+      .locator("#description-inline-expander:visible")
+      .textContent();
+    console.log("Description text:", descriptionText?.trim());
+
+    await page
+      .locator("#description-inline-expander:visible")
+      .scrollIntoViewIfNeeded();
+    // Check that the description contains the original English title as fallback and not the Russian title translation fallback
+    expect(descriptionText).toContain("Answer The Call, Win $10,000");
+    expect(descriptionText).not.toContain("Ответь На Звонок, Выиграй $10,000");
+
+    // Take a screenshot for visual verification
+    await page.screenshot({
+      path: `images/tests/${browserNameWithExtensions}/${localeString}/youtube-extension-test-description-fallback.png`,
     });
 
     // Check console message count
@@ -362,7 +424,8 @@ test.describe("YouTube Anti-Translate extension", () => {
     );
 
     // Wait for the shorts title element to be present
-    const shortsTitleSelector = "yt-shorts-video-title-view-model > h2 > span";
+    const shortsTitleSelector =
+      "#yt-anti-translate-fake-node-current-short-video";
     await page.waitForSelector(shortsTitleSelector);
 
     // Get the title text
@@ -445,7 +508,9 @@ test.describe("YouTube Anti-Translate extension", () => {
       await originalVideo.scrollIntoViewIfNeeded();
       try {
         await page.waitForLoadState("networkidle", { timeout: 5000 });
-      } catch {}
+      } catch {
+        // empty
+      }
     }
     const translatedVideo = page.locator(translatedVideoSelector).first();
     if (await translatedVideo.isVisible()) {
@@ -453,7 +518,9 @@ test.describe("YouTube Anti-Translate extension", () => {
       await translatedVideo.scrollIntoViewIfNeeded();
       try {
         await page.waitForLoadState("networkidle", { timeout: 5000 });
-      } catch {}
+      } catch {
+        // empty
+      }
     }
 
     console.log("Checking Videos tab for original title...");
@@ -466,7 +533,9 @@ test.describe("YouTube Anti-Translate extension", () => {
     await page.locator("#tabsContent").getByText("Shorts").click();
     try {
       await page.waitForLoadState("networkidle", { timeout: 5000 });
-    } catch {}
+    } catch {
+      // empty
+    }
     await page.waitForTimeout(1000); // Give it a moment to load more items if needed
 
     // --- Check Shorts Tab ---
@@ -483,7 +552,9 @@ test.describe("YouTube Anti-Translate extension", () => {
       await originalShort.scrollIntoViewIfNeeded();
       try {
         await page.waitForLoadState("networkidle", { timeout: 5000 });
-      } catch {}
+      } catch {
+        // empty
+      }
     }
     const translatedShort = page.locator(translatedShortSelector).first();
     if (await translatedShort.isVisible()) {
@@ -491,7 +562,9 @@ test.describe("YouTube Anti-Translate extension", () => {
       await translatedShort.scrollIntoViewIfNeeded();
       try {
         await page.waitForLoadState("networkidle", { timeout: 5000 });
-      } catch {}
+      } catch {
+        // empty
+      }
     }
     await page.waitForTimeout(1000); // Give it a moment to load more items if needed
 
@@ -506,7 +579,9 @@ test.describe("YouTube Anti-Translate extension", () => {
     await page.locator("#tabsContent").getByText("Видео").click();
     try {
       await page.waitForLoadState("networkidle", { timeout: 5000 });
-    } catch {}
+    } catch {
+      // empty
+    }
     await page.waitForSelector(
       "ytd-rich-grid-media >> ytd-thumbnail-overlay-time-status-renderer:not([overlay-style='SHORTS'])",
       { state: "visible" },
@@ -562,7 +637,9 @@ test.describe("YouTube Anti-Translate extension", () => {
       await firstShort.click();
       try {
         await page.waitForLoadState("networkidle", { timeout: 5000 });
-      } catch {}
+      } catch {
+        // empty
+      }
     }
     await page.waitForTimeout(2000);
 
@@ -620,7 +697,9 @@ test.describe("YouTube Anti-Translate extension", () => {
       await buttonDown.click();
       try {
         await page.waitForLoadState("networkidle", { timeout: 5000 });
-      } catch {}
+      } catch {
+        // empty
+      }
     }
     await page.waitForTimeout(2000);
 
@@ -659,7 +738,9 @@ test.describe("YouTube Anti-Translate extension", () => {
       await buttonDown2.click();
       try {
         await page.waitForLoadState("networkidle", { timeout: 5000 });
-      } catch {}
+      } catch {
+        // empty
+      }
     }
     await page.waitForTimeout(2000);
 
@@ -699,10 +780,14 @@ test.describe("YouTube Anti-Translate extension", () => {
 
     /**
      * If Track name is "Default" that is always an advert
-     * @param currentTrack the audio track that could be of an advert
+     * @param currentTrack - the audio track that could be of an advert
+     * @param currentVideoId - the video id that could be of an advert
      * @returns a new short audio track and video id
      */
-    async function IfAdvertThenReturnNext(currentTrack, currentVideoId) {
+    async function IfAdvertThenReturnNext(
+      currentTrack: any,
+      currentVideoId: string | null,
+    ) {
       if (
         currentTrack[getTrackLanguageFieldObjectName(currentTrack)!]?.name ===
         "Default"
@@ -715,7 +800,9 @@ test.describe("YouTube Anti-Translate extension", () => {
           await buttonDown2.click();
           try {
             await page.waitForLoadState("networkidle", { timeout: 5000 });
-          } catch {}
+          } catch {
+            // empty
+          }
         }
         await page.waitForTimeout(2000);
 
@@ -866,7 +953,20 @@ test.describe("YouTube Anti-Translate extension", () => {
     await page.goto("https://www.youtube.com/embed/iLU0CE2c2HQ");
 
     // Wait for the video to load
-    await page.waitForSelector("video", { timeout: 10000 });
+    try {
+      await page.waitForSelector("video", { timeout: 10000 });
+    } catch {
+      await page.waitForSelector(
+        "div.ytp-error-content-wrap-subreason > span:has-text('153')",
+        { timeout: 10000 },
+      );
+      // If we hit a 153 error (playback not available) then skip the rest of the test
+      console.log(
+        "Video playback not available, skipping the rest of the test.",
+      );
+      await context.close();
+      testInfo.skip();
+    }
 
     await page.click("#movie_player > div.ytp-cued-thumbnail-overlay > button");